src/pyproject.toml
[project]
name = "samples"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests>=2.32.0",
    "azure-identity>=1.17.0",
    "azure-ai-agents>=1.0.0b6",
    "azure-ai-projects>=1.0.0b3",
    "types-requests>=2.32.0.20240914",
    "azure-mgmt-authorization>=4.0.0",
    "azure-mgmt-machinelearningservices>=2.0.0b2",
]
